+++ examples/rotating-cube.mjs
@@ -120,22 +120,22 @@ const fsSrc = `
   });
 
   const stagedVertexBuffer = device.createBuffer({
-    size: BigInt(modelVertices.byteLength),
+    size: modelVertices.byteLength,
     usage: GPUBufferUsage.VERTEX | GPUBufferUsage.COPY_DST
   });
-  stagedVertexBuffer.setSubData(0n, modelVertices);
+  stagedVertexBuffer.setSubData(0, modelVertices);
 
   const stagedIndexBuffer = device.createBuffer({
-    size: BigInt(modelIndices.byteLength),
+    size: modelIndices.byteLength,
     usage: GPUBufferUsage.INDEX | GPUBufferUsage.COPY_DST
   });
-  stagedIndexBuffer.setSubData(0n, modelIndices);
+  stagedIndexBuffer.setSubData(0, modelIndices);
 
   const stagedUniformBuffer = device.createBuffer({
-    size: BigInt(mModelViewProjection.byteLength),
+    size: mModelViewProjection.byteLength,
     usage: GPUBufferUsage.UNIFORM | GPUBufferUsage.COPY_DST
   });
-  stagedUniformBuffer.setSubData(0n, mModelViewProjection);
+  stagedUniformBuffer.setSubData(0, mModelViewProjection);
 
   const uniformBindGroupLayout = device.createBindGroupLayout({
     bindings: [{
@@ -176,17 +176,17 @@ const fsSrc = `
       indexFormat: "uint32",
       vertexBuffers: [
         {
-          arrayStride: BigInt(6 * Float32Array.BYTES_PER_ELEMENT),
+          arrayStride: 6 * Float32Array.BYTES_PER_ELEMENT,
           stepMode: "vertex",
           attributes: [
             {
               shaderLocation: 0,
-              offset: BigInt(0 * Float32Array.BYTES_PER_ELEMENT),
+              offset: 0 * Float32Array.BYTES_PER_ELEMENT,
               format: "float3"
             },
             {
               shaderLocation: 1,
-              offset: BigInt(3 * Float32Array.BYTES_PER_ELEMENT),
+              offset: 3 * Float32Array.BYTES_PER_ELEMENT,
               format: "float3"
             }
           ]
@@ -209,8 +209,8 @@ const fsSrc = `
     bindings: [{
       binding: 0,
       buffer: stagedUniformBuffer,
-      offset: 0n,
-      size: BigInt(mModelViewProjection.byteLength)
+      offset: 0,
+      size: mModelViewProjection.byteLength
     }]
   });
 
@@ -247,7 +247,7 @@ const fsSrc = `
     mat4.scale(mModel, mModel, vec3.fromValues(3, 3, 3));
     mat4.multiply(mModelViewProjection, mView, mModel);
     mat4.multiply(mModelViewProjection, mProjection, mModelViewProjection);
-    stagedUniformBuffer.setSubData(0n, mModelViewProjection);
+    stagedUniformBuffer.setSubData(0, mModelViewProjection);
 
     const backBufferView = swapChain.getCurrentTextureView();
     const commandEncoder = device.createCommandEncoder({});
